+++ examples/basic.rs
@@ -1,6 +1,4 @@
-use redis;
-use redis::{transaction, Commands, PipelineCommands};
-use std::error::Error;
+use redis::{self, transaction, Commands};
 
 use std::collections::HashMap;
 
@@ -159,7 +157,7 @@ fn main() {
     match do_redis_code() {
         Err(err) => {
             println!("Could not execute example:");
-            println!("  {}: {}", err.category(), err.description());
+            println!("  {}: {}", err.category(), err);
         }
         Ok(()) => {}
     }