+++ src/commands.rs
@@ -1,6 +1,6 @@
 // can't use rustfmt here because it screws up the file.
 #![cfg_attr(rustfmt, rustfmt_skip)]
-use crate::types::{FromRedisValue, ToRedisArgs, RedisResult, NumericBehavior};
+use crate::types::{FromRedisValue, ToRedisArgs, RedisResult, RedisFuture, NumericBehavior};
 use crate::connection::{ConnectionLike, Msg, Connection};
 use crate::cmd::{cmd, Cmd, Pipeline, Iter};
 
@@ -9,6 +9,7 @@ use crate::geo;
 
 macro_rules! implement_commands {
     (
+        $lifetime: lifetime
         $(
             $(#[$attr:meta])+
             fn $name:ident<$($tyargs:ident : $ty:ident),*>(
@@ -40,6 +41,7 @@ macro_rules! implement_commands {
         /// use redis::Commands;
         /// let client = redis::Client::open("redis://127.0.0.1/")?;
         /// let mut con = client.get_connection()?;
+        /// let () = con.set("my_key", 42)?;
         /// assert_eq!(con.get("my_key"), Ok(42));
         /// # Ok(()) }
         /// ```
@@ -47,9 +49,9 @@ macro_rules! implement_commands {
             $(
                 $(#[$attr])*
                 #[inline]
-                fn $name<$($tyargs: $ty,)* RV: FromRedisValue>(
+                fn $name<$lifetime, $($tyargs: $ty, )* RV: FromRedisValue>(
                     &mut self $(, $argname: $argty)*) -> RedisResult<RV>
-                    { ($body).query(self) }
+                    { Cmd::$name($($argname),*).query(self) }
             )*
 
             /// Incrementally iterate the keys space.
@@ -121,25 +123,75 @@ macro_rules! implement_commands {
             }
         }
 
-        /// Implements common redis commands for pipelines.  Unlike the regular
-        /// commands trait, this returns the pipeline rather than a result
-        /// directly.  Other than that it works the same however.
-        pub trait PipelineCommands {
-            #[doc(hidden)]
-            fn perform(&mut self, con: Cmd) -> &mut Self;
+        impl Cmd {
+            $(
+                $(#[$attr])*
+                pub fn $name<$lifetime, $($tyargs: $ty),*>($($argname: $argty),*) -> Self {
+                    ::std::mem::replace($body, Cmd::new())
+                }
+            )*
+        }
 
+        /// Implements common redis commands over asynchronous connections. This
+        /// allows you to send commands straight to a connection or client.
+        ///
+        /// This allows you to use nicer syntax for some common operations.
+        /// For instance this code:
+        ///
+        /// ```rust,no_run
+        /// use redis::AsyncCommands;
+        /// # async fn do_something() -> redis::RedisResult<()> {
+        /// let client = redis::Client::open("redis://127.0.0.1/")?;
+        /// let mut con = client.get_async_connection().await?;
+        /// let () = redis::cmd("SET").arg("my_key").arg(42i32).query_async(&mut con).await?;
+        /// assert_eq!(redis::cmd("GET").arg("my_key").query_async(&mut con).await, Ok(42i32));
+        /// # Ok(()) }
+        /// ```
+        ///
+        /// Will become this:
+        ///
+        /// ```rust,no_run
+        /// use redis::AsyncCommands;
+        /// # async fn do_something() -> redis::RedisResult<()> {
+        /// use redis::Commands;
+        /// let client = redis::Client::open("redis://127.0.0.1/")?;
+        /// let mut con = client.get_async_connection().await?;
+        /// let () = con.set("my_key", 42i32).await?;
+        /// assert_eq!(con.get("my_key").await, Ok(42i32));
+        /// # Ok(()) }
+        /// ```
+        pub trait AsyncCommands : crate::aio::ConnectionLike + Send + Sized {
             $(
                 $(#[$attr])*
                 #[inline]
-                fn $name<$($tyargs: $ty),*>(
+                fn $name<$lifetime, $($tyargs: $ty + Send + Sync + $lifetime,)* RV>(
+                    & $lifetime mut self
+                    $(, $argname: $argty)*
+                    ) -> RedisFuture<'a, RV>
+                where RV: FromRedisValue,
+                {
+                    Box::pin(async move { ($body).query_async(self).await })
+                }
+            )*
+        }
+
+        /// Implements common redis commands for pipelines.  Unlike the regular
+        /// commands trait, this returns the pipeline rather than a result
+        /// directly.  Other than that it works the same however.
+        impl Pipeline {
+            $(
+                $(#[$attr])*
+                #[inline]
+                pub fn $name<$lifetime, $($tyargs: $ty),*>(
                     &mut self $(, $argname: $argty)*) -> &mut Self
-                    { self.perform(::std::mem::replace($body, Cmd::new())) }
+                    { self.add_command(::std::mem::replace($body, Cmd::new())) }
             )*
         }
     )
 }
 
 implement_commands! {
+    'a
     // most common operations
 
     /// Get the value of a key.  If key is a vec this becomes an `MGET`.
@@ -158,7 +210,7 @@ implement_commands! {
     }
 
     /// Sets multiple keys to their values.
-    fn set_multiple<K: ToRedisArgs, V: ToRedisArgs>(items: &[(K, V)]) {
+    fn set_multiple<K: ToRedisArgs, V: ToRedisArgs>(items: &'a [(K, V)]) {
         cmd("MSET").arg(items)
     }
 
@@ -167,13 +219,18 @@ implement_commands! {
         cmd("SETEX").arg(key).arg(seconds).arg(value)
     }
 
+    /// Set the value and expiration in milliseconds of a key.
+    fn pset_ex<K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V, milliseconds: usize) {
+        cmd("PSETEX").arg(key).arg(milliseconds).arg(value)
+    }
+
     /// Set the value of a key, only if the key does not exist
     fn set_nx<K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) {
         cmd("SETNX").arg(key).arg(value)
     }
 
     /// Sets multiple keys to their values failing if at least one already exists.
-    fn mset_nx<K: ToRedisArgs, V: ToRedisArgs>(items: &[(K, V)]) {
+    fn mset_nx<K: ToRedisArgs, V: ToRedisArgs>(items: &'a [(K, V)]) {
         cmd("MSETNX").arg(items)
     }
 
@@ -227,9 +284,14 @@ implement_commands! {
         cmd("PERSIST").arg(key)
     }
 
-    /// Check the expiration time of a key.
+    /// Get the expiration time of a key.
     fn ttl<K: ToRedisArgs>(key: K) {
         cmd("TTL").arg(key)
+    }
+
+    /// Get the expiration time of a key in milliseconds.
+    fn pttl<K: ToRedisArgs>(key: K) {
+        cmd("PTTL").arg(key)
     }
 
     /// Rename a key.
@@ -331,7 +393,7 @@ implement_commands! {
     }
 
     /// Sets a multiple fields in a hash.
-    fn hset_multiple<K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(key: K, items: &[(F, V)]) {
+    fn hset_multiple<K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(key: K, items: &'a [(F, V)]) {
         cmd("HMSET").arg(key).arg(items)
     }
 
@@ -553,7 +615,7 @@ implement_commands! {
     }
 
     /// Add multiple members to a sorted set, or update its score if it already exists.
-    fn zadd_multiple<K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(key: K, items: &[(S, M)]) {
+    fn zadd_multiple<K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(key: K, items: &'a [(S, M)]) {
         cmd("ZADD").arg(key).arg(items)
     }
 
@@ -575,19 +637,19 @@ implement_commands! {
 
     /// Intersect multiple sorted sets and store the resulting sorted set in
     /// a new key using SUM as aggregation function.
-    fn zinterstore<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zinterstore<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZINTERSTORE").arg(dstkey).arg(keys.len()).arg(keys)
     }
 
     /// Intersect multiple sorted sets and store the resulting sorted set in
     /// a new key using MIN as aggregation function.
-    fn zinterstore_min<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zinterstore_min<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZINTERSTORE").arg(dstkey).arg(keys.len()).arg(keys).arg("AGGREGATE").arg("MIN")
     }
 
     /// Intersect multiple sorted sets and store the resulting sorted set in
     /// a new key using MAX as aggregation function.
-    fn zinterstore_max<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zinterstore_max<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZINTERSTORE").arg(dstkey).arg(keys.len()).arg(keys).arg("AGGREGATE").arg("MAX")
     }
 
@@ -669,8 +731,8 @@ implement_commands! {
     }
 
     /// Remove all members in a sorted set within the given indexes.
-    fn zrembyrank<K: ToRedisArgs>(key: K, start: isize, stop: isize) {
-        cmd("ZREMBYRANK").arg(key).arg(start).arg(stop)
+    fn zremrangebyrank<K: ToRedisArgs>(key: K, start: isize, stop: isize) {
+        cmd("ZREMRANGEBYRANK").arg(key).arg(start).arg(stop)
     }
 
     /// Remove all members in a sorted set within the given scores.
@@ -725,19 +787,19 @@ implement_commands! {
 
     /// Unions multiple sorted sets and store the resulting sorted set in
     /// a new key using SUM as aggregation function.
-    fn zunionstore<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zunionstore<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZUNIONSTORE").arg(dstkey).arg(keys.len()).arg(keys)
     }
 
     /// Unions multiple sorted sets and store the resulting sorted set in
     /// a new key using MIN as aggregation function.
-    fn zunionstore_min<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zunionstore_min<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZUNIONSTORE").arg(dstkey).arg(keys.len()).arg(keys).arg("AGGREGATE").arg("MIN")
     }
 
     /// Unions multiple sorted sets and store the resulting sorted set in
     /// a new key using MAX as aggregation function.
-    fn zunionstore_max<K: ToRedisArgs>(dstkey: K, keys: &[K]) {
+    fn zunionstore_max<K: ToRedisArgs>(dstkey: K, keys: &'a [K]) {
         cmd("ZUNIONSTORE").arg(dstkey).arg(keys.len()).arg(keys).arg("AGGREGATE").arg("MAX")
     }
 
@@ -800,6 +862,7 @@ implement_commands! {
     /// }
     /// ```
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_add<K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) {
         cmd("GEOADD").arg(key).arg(members)
     }
@@ -836,6 +899,7 @@ implement_commands! {
     /// }
     /// ```
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_dist<K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs>(
         key: K,
         member1: M1,
@@ -869,6 +933,7 @@ implement_commands! {
     /// }
     /// ```
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_hash<K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) {
         cmd("GEOHASH").arg(key).arg(members)
     }
@@ -897,6 +962,7 @@ implement_commands! {
     /// }
     /// ```
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_pos<K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) {
         cmd("GEOPOS").arg(key).arg(members)
     }
@@ -921,6 +987,7 @@ implement_commands! {
     /// }
     /// ```
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_radius<K: ToRedisArgs>(
         key: K,
         longitude: f64,
@@ -941,6 +1008,7 @@ implement_commands! {
     /// Retrieve members selected by distance with the center of `member`. The
     /// member itself is always contained in the results.
     #[cfg(feature = "geospatial")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
     fn geo_radius_by_member<K: ToRedisArgs, M: ToRedisArgs>(
         key: K,
         member: M,
@@ -1026,6 +1094,8 @@ pub trait PubSubCommands: Sized {
 
 impl<T> Commands for T where T: ConnectionLike {}
 
+impl<T> AsyncCommands for T where T: crate::aio::ConnectionLike + Send + ?Sized {}
+
 impl PubSubCommands for Connection {
     fn subscribe<C, F, U>(&mut self, channels: C, mut func: F) -> RedisResult<U>
         where F: FnMut(Msg) -> ControlFlow<U>,
@@ -1057,11 +1127,5 @@ impl PubSubCommands for Connection {
                 ControlFlow::Break(value) => return Ok(value),
             }
         };
-    }
-}
-
-impl PipelineCommands for Pipeline {
-    fn perform(&mut self, cmd: Cmd) -> &mut Pipeline {
-        self.add_command(cmd)
     }
 }