+++ src/lib.rs
@@ -205,12 +205,10 @@
 //! # Ok(()) }
 //! ```
 //!
-//! You can also use high-level commands on pipelines through the
-//! `PipelineCommands` trait:
+//! You can also use high-level commands on pipelines:
 //!
 //! ```rust,no_run
 //! # fn do_something() -> redis::RedisResult<()> {
-//! use redis::PipelineCommands;
 //! # let client = redis::Client::open("redis://127.0.0.1/").unwrap();
 //! # let mut con = client.get_connection().unwrap();
 //! let (k1, k2) : (i32, i32) = redis::pipe()
@@ -230,7 +228,7 @@
 //!
 //! ```rust,no_run
 //! # fn do_something() -> redis::RedisResult<()> {
-//! use redis::{Commands, PipelineCommands};
+//! use redis::Commands;
 //! # let client = redis::Client::open("redis://127.0.0.1/").unwrap();
 //! # let mut con = client.get_connection().unwrap();
 //! let key = "the_key";
@@ -300,38 +298,24 @@
 //! with a few concessions to make it fit the constraints of `futures`.
 //!
 //! ```rust,no_run
-//! extern crate redis;
-//! extern crate futures;
-//! extern crate tokio;
+//! use futures::prelude::*;
+//! use redis::AsyncCommands;
 //!
-//! use futures::Future;
-//!
-//! # fn main() {
+//! # #[tokio::main]
+//! # async fn main() -> redis::RedisResult<()> {
 //! let client = redis::Client::open("redis://127.0.0.1/").unwrap();
-//! let connect = client.get_async_connection();
+//! let mut con = client.get_async_connection().await?;
 //!
-//! tokio::run(connect.and_then(|con| {
-//!     redis::cmd("SET")
-//!         .arg("key1")
-//!         .arg(b"foo")
-//!         // `query_async` acts in the same way as `query` but requires the connection to be
-//!         // taken by value as the method returns a `Future` instead of `Result`.
-//!         // This connection will be returned after the future has been completed allowing it to
-//!         // be used again.
-//!         .query_async(con)
-//!         .and_then(|(con, ())| {
-//!             redis::cmd("SET").arg(&["key2", "bar"]).query_async(con)
-//!         })
-//!         .and_then(|(con, ())| {
-//!             redis::cmd("MGET")
-//!                 .arg(&["key1", "key2"])
-//!                 .query_async(con)
-//!                 .map(|t| t.1)
-//!         })
-//! }).then(|result| {
-//!     assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
-//!     Ok(())
-//! }));
+//! let () = con.set("key1", b"foo").await?;
+//!
+//! let () = redis::cmd("SET").arg(&["key2", "bar"]).query_async(&mut con).await?;
+//!
+//! let result = redis::cmd("MGET")
+//!     .arg(&["key1", "key2"])
+//!     .query_async(&mut con)
+//!     .await;
+//! assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
+//! Ok(())
 //! # }
 //! ```
 //!
@@ -350,8 +334,8 @@
 
 // public api
 pub use crate::client::Client;
-pub use crate::cmd::{cmd, pack_command, pipe, Cmd, Iter, Pipeline};
-pub use crate::commands::{Commands, ControlFlow, PipelineCommands, PubSubCommands};
+pub use crate::cmd::{cmd, pack_command, pipe, Arg, Cmd, Iter, Pipeline};
+pub use crate::commands::{AsyncCommands, Commands, ControlFlow, PubSubCommands};
 pub use crate::connection::{
     parse_redis_url, transaction, Connection, ConnectionAddr, ConnectionInfo, ConnectionLike,
     IntoConnectionInfo, Msg, PubSub,
@@ -389,6 +373,7 @@ mod macros;
 pub mod aio;
 
 #[cfg(feature = "geospatial")]
+#[cfg_attr(docsrs, doc(cfg(feature = "geospatial")))]
 pub mod geo;
 
 #[cfg(feature = "cluster")]