+++ src/script.rs
@@ -1,10 +1,9 @@
 use sha1::Sha1;
 
 use crate::aio;
-use crate::cmd::{cmd, Cmd};
+use crate::cmd::cmd;
 use crate::connection::ConnectionLike;
-use crate::types::{ErrorKind, FromRedisValue, RedisError, RedisFuture, RedisResult, ToRedisArgs};
-use futures::{try_ready, Async, Future, Poll};
+use crate::types::{ErrorKind, FromRedisValue, RedisResult, ToRedisArgs};
 
 /// Represents a lua script.
 pub struct Script {
@@ -151,10 +150,10 @@ impl<'a> ScriptInvocation<'a> {
 
     /// Asynchronously invokes the script and returns the result.
     #[inline]
-    pub fn invoke_async<C, T>(&self, con: C) -> impl Future<Item = (C, T), Error = RedisError>
+    pub async fn invoke_async<C, T>(&self, con: &mut C) -> RedisResult<T>
     where
-        C: aio::ConnectionLike + Clone + Send + 'static,
-        T: FromRedisValue + Send + 'static,
+        C: aio::ConnectionLike,
+        T: FromRedisValue,
     {
         let mut eval_cmd = cmd("EVALSHA");
         eval_cmd
@@ -165,63 +164,20 @@ impl<'a> ScriptInvocation<'a> {
 
         let mut load_cmd = cmd("SCRIPT");
         load_cmd.arg("LOAD").arg(self.script.code.as_bytes());
-
-        let future = eval_cmd.query_async(con.clone());
-        InvokeAsyncFuture {
-            con: con.clone(),
-            eval_cmd,
-            load_cmd,
-            future: CmdFuture::Eval(future),
-        }
-    }
-}
-
-/// A future that runs the given script and loads it into Redis if
-/// it has not already been loaded
-struct InvokeAsyncFuture<C, T> {
-    con: C,
-    eval_cmd: Cmd,
-    load_cmd: Cmd,
-    future: CmdFuture<C, T>,
-}
-
-enum CmdFuture<C, T> {
-    Load(RedisFuture<(C, String)>),
-    Eval(RedisFuture<(C, T)>),
-}
-
-impl<C, T> Future for InvokeAsyncFuture<C, T>
-where
-    C: aio::ConnectionLike + Clone + Send + 'static,
-    T: FromRedisValue + Send + 'static,
-{
-    type Item = (C, T);
-    type Error = RedisError;
-
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        loop {
-            self.future = match self.future {
-                CmdFuture::Load(ref mut future) => {
-                    // When we're done loading the script into Redis, try eval'ing it again
-                    let (con, _hash) = try_ready!(future.poll());
-                    CmdFuture::Eval(self.eval_cmd.query_async(con))
+        match eval_cmd.query_async(con).await {
+            Ok(val) => {
+                // Return the value from the script evaluation
+                return Ok(val).into();
+            }
+            Err(err) => {
+                // Load the script into Redis if the script hash wasn't there already
+                if err.kind() == ErrorKind::NoScriptError {
+                    let _hash = load_cmd.query_async(con).await?;
+                    eval_cmd.query_async(con).await
+                } else {
+                    Err(err)
                 }
-                CmdFuture::Eval(ref mut future) => match future.poll() {
-                    Ok(Async::NotReady) => return Ok(Async::NotReady),
-                    Ok(Async::Ready((con, val))) => {
-                        // Return the value from the script evaluation
-                        return Ok(Async::Ready((con, val)));
-                    }
-                    Err(err) => {
-                        // Load the script into Redis if the script hash wasn't there already
-                        if err.kind() == ErrorKind::NoScriptError {
-                            CmdFuture::Load(self.load_cmd.query_async(self.con.clone()))
-                        } else {
-                            return Err(err);
-                        }
-                    }
-                },
-            };
+            }
         }
     }
 }