+++ tests/test_async.rs
@@ -2,14 +2,11 @@ use redis;
 
 use futures;
 
-use futures::{future, Future};
+use futures::{future, prelude::*};
 
 use crate::support::*;
 
-use redis::aio::SharedConnection;
-use redis::RedisError;
-
-use tokio::runtime::current_thread::block_on_all;
+use redis::{aio::MultiplexedConnection, RedisError, RedisResult};
 
 mod support;
 
@@ -18,69 +15,75 @@ fn test_args() {
     let ctx = TestContext::new();
     let connect = ctx.async_connection();
 
-    block_on_all(connect.and_then(|con| {
-        redis::cmd("SET")
-            .arg("key1")
-            .arg(b"foo")
-            .query_async(con)
-            .and_then(|(con, ())| redis::cmd("SET").arg(&["key2", "bar"]).query_async(con))
-            .and_then(|(con, ())| {
-                redis::cmd("MGET")
-                    .arg(&["key1", "key2"])
-                    .query_async(con)
-                    .map(|t| t.1)
-            })
-            .then(|result| {
-                assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
-                result
-            })
+    block_on_all(connect.and_then(|mut con| {
+        async move {
+            let () = redis::cmd("SET")
+                .arg("key1")
+                .arg(b"foo")
+                .query_async(&mut con)
+                .await?;
+            let () = redis::cmd("SET")
+                .arg(&["key2", "bar"])
+                .query_async(&mut con)
+                .await?;
+            let result = redis::cmd("MGET")
+                .arg(&["key1", "key2"])
+                .query_async(&mut con)
+                .await;
+            assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
+            result
+        }
     }))
     .unwrap();
 }
 
 #[test]
-fn dont_panic_on_closed_shared_connection() {
+fn dont_panic_on_closed_multiplexed_connection() {
     let ctx = TestContext::new();
-    let connect = ctx.shared_async_connection();
+    let connect = ctx.multiplexed_async_connection();
     drop(ctx);
 
-    block_on_all(future::lazy(|| {
+    block_on_all(async move {
         connect
             .and_then(|con| {
-                let cmd = move || {
-                    redis::cmd("SET")
-                        .arg("key1")
-                        .arg(b"foo")
-                        .query_async(con.clone())
-                        .map(|(_, ())| ())
-                };
-                cmd().then(move |result| {
+                async move {
+                    let cmd = move || {
+                        let mut con = con.clone();
+                        async move {
+                            redis::cmd("SET")
+                                .arg("key1")
+                                .arg(b"foo")
+                                .query_async(&mut con)
+                                .await
+                        }
+                    };
+                    let result: RedisResult<()> = cmd().await;
                     assert_eq!(
                         result.as_ref().unwrap_err().kind(),
                         redis::ErrorKind::IoError,
                         "{}",
                         result.as_ref().unwrap_err()
                     );
-                    cmd()
-                })
+                    cmd().await
+                }
             })
-            .then(|result| -> Result<(), ()> {
+            .map(|result| {
                 assert_eq!(
                     result.as_ref().unwrap_err().kind(),
                     redis::ErrorKind::IoError,
                     "{}",
                     result.as_ref().unwrap_err()
                 );
-                Ok(())
             })
-    }))
-    .unwrap();
+            .await
+    });
 }
 
 #[test]
 fn test_pipeline_transaction() {
     let ctx = TestContext::new();
-    block_on_all(ctx.async_connection().and_then(|con| {
+    block_on_all(async move {
+        let mut con = ctx.async_connection().await?;
         let mut pipe = redis::pipe();
         pipe.atomic()
             .cmd("SET")
@@ -93,134 +96,142 @@ fn test_pipeline_transaction() {
             .ignore()
             .cmd("MGET")
             .arg(&["key_1", "key_2"]);
-        pipe.query_async(con)
-            .and_then(|(_con, ((k1, k2),)): (_, ((i32, i32),))| {
+        pipe.query_async(&mut con)
+            .map_ok(|((k1, k2),): ((i32, i32),)| {
                 assert_eq!(k1, 42);
                 assert_eq!(k2, 43);
-                Ok(())
             })
-    }))
+            .await
+    })
     .unwrap();
 }
 
-fn test_cmd(
-    con: &SharedConnection,
-    i: i32,
-) -> Box<dyn Future<Item = (), Error = RedisError> + Send> {
-    let key = format!("key{}", i);
-    let key_2 = key.clone();
-    let key2 = format!("key{}_2", i);
-    let key2_2 = key2.clone();
+fn test_cmd(con: &MultiplexedConnection, i: i32) -> impl Future<Output = RedisResult<()>> + Send {
+    let mut con = con.clone();
+    async move {
+        let key = format!("key{}", i);
+        let key_2 = key.clone();
+        let key2 = format!("key{}_2", i);
+        let key2_2 = key2.clone();
 
-    let foo_val = format!("foo{}", i);
+        let foo_val = format!("foo{}", i);
 
-    let con1 = con.clone();
-    let con2 = con.clone();
-    Box::new(
-        redis::cmd("SET")
+        let () = redis::cmd("SET")
             .arg(&key[..])
             .arg(foo_val.as_bytes())
-            .query_async(con.clone())
-            .and_then(move |(_, ())| redis::cmd("SET").arg(&[&key2, "bar"]).query_async(con1))
-            .and_then(move |(_, ())| {
-                redis::cmd("MGET")
-                    .arg(&[&key_2, &key2_2])
-                    .query_async(con2)
-                    .map(|t| t.1)
-                    .then(|result| {
-                        assert_eq!(Ok((foo_val, b"bar".to_vec())), result);
-                        Ok(())
-                    })
-            }),
-    )
-}
-
-fn test_error(con: &SharedConnection) -> Box<dyn Future<Item = (), Error = RedisError> + Send> {
-    Box::new(
-        redis::cmd("SET")
-            .query_async(con.clone())
-            .then(|result| match result {
-                Ok((_, ())) => panic!("Expected redis to return an error"),
-                Err(_) => Ok(()),
-            }),
-    )
-}
-
-#[test]
-fn test_args_shared_connection() {
-    let ctx = TestContext::new();
-    block_on_all(future::lazy(|| {
-        ctx.shared_async_connection()
-            .and_then(|con| {
-                let cmds = (0..100).map(move |i| test_cmd(&con, i));
-                future::join_all(cmds).map(|results| {
-                    assert_eq!(results.len(), 100);
-                })
-            })
-            .map_err(|err| panic!("{}", err))
-    }))
-    .unwrap();
-}
-
-#[test]
-fn test_args_with_errors_shared_connection() {
-    let ctx = TestContext::new();
-    block_on_all(future::lazy(|| {
-        ctx.shared_async_connection()
-            .and_then(|con| {
-                let cmds = (0..100).map(move |i| {
-                    if i % 2 == 0 {
-                        test_cmd(&con, i)
-                    } else {
-                        test_error(&con)
-                    }
-                });
-                future::join_all(cmds).map(|results| {
-                    assert_eq!(results.len(), 100);
-                })
-            })
-            .map_err(|err| panic!("{}", err))
-    }))
-    .unwrap();
-}
-
-#[test]
-fn test_transaction_shared_connection() {
-    let ctx = TestContext::new();
-    block_on_all(future::lazy(|| {
-        ctx.shared_async_connection()
-            .and_then(|con| {
-                let cmds = (0..100).map(move |i| {
-                    let foo_val = i;
-                    let bar_val = format!("bar{}", i);
-
-                    let mut pipe = redis::pipe();
-                    pipe.atomic()
-                        .cmd("SET")
-                        .arg("key")
-                        .arg(foo_val)
-                        .ignore()
-                        .cmd("SET")
-                        .arg(&["key2", &bar_val[..]])
-                        .ignore()
-                        .cmd("MGET")
-                        .arg(&["key", "key2"]);
-
-                    pipe.query_async(con.clone())
-                        .map(|t| t.1)
-                        .then(move |result| {
-                            assert_eq!(Ok(((foo_val, bar_val.clone().into_bytes()),)), result);
-                            result
-                        })
-                });
-                future::join_all(cmds)
-            })
-            .and_then(|results| {
-                assert_eq!(results.len(), 100);
+            .query_async(&mut con)
+            .await?;
+        let () = redis::cmd("SET")
+            .arg(&[&key2, "bar"])
+            .query_async(&mut con)
+            .await?;
+        redis::cmd("MGET")
+            .arg(&[&key_2, &key2_2])
+            .query_async(&mut con)
+            .map(|result| {
+                assert_eq!(Ok((foo_val, b"bar".to_vec())), result);
                 Ok(())
             })
+            .await
+    }
+}
+
+fn test_error(con: &MultiplexedConnection) -> impl Future<Output = RedisResult<()>> {
+    let mut con = con.clone();
+    async move {
+        redis::cmd("SET")
+            .query_async(&mut con)
+            .map(|result| match result {
+                Ok(()) => panic!("Expected redis to return an error"),
+                Err(_) => Ok(()),
+            })
+            .await
+    }
+}
+
+#[test]
+fn test_args_multiplexed_connection() {
+    let ctx = TestContext::new();
+    block_on_all(async move {
+        ctx.multiplexed_async_connection()
+            .and_then(|con| {
+                let cmds = (0..100).map(move |i| test_cmd(&con, i));
+                future::try_join_all(cmds).map_ok(|results| {
+                    assert_eq!(results.len(), 100);
+                })
+            })
             .map_err(|err| panic!("{}", err))
-    }))
+            .await
+    })
+    .unwrap();
+}
+
+#[test]
+fn test_args_with_errors_multiplexed_connection() {
+    let ctx = TestContext::new();
+    block_on_all(async move {
+        ctx.multiplexed_async_connection()
+            .and_then(|con| {
+                let cmds = (0..100).map(move |i| {
+                    let con = con.clone();
+                    async move {
+                        if i % 2 == 0 {
+                            test_cmd(&con, i).await
+                        } else {
+                            test_error(&con).await
+                        }
+                    }
+                });
+                future::try_join_all(cmds).map_ok(|results| {
+                    assert_eq!(results.len(), 100);
+                })
+            })
+            .map_err(|err| panic!("{}", err))
+            .await
+    })
+    .unwrap();
+}
+
+#[test]
+fn test_transaction_multiplexed_connection() {
+    let ctx = TestContext::new();
+    block_on_all(async move {
+        ctx.multiplexed_async_connection()
+            .and_then(|con| {
+                let cmds = (0..100).map(move |i| {
+                    let mut con = con.clone();
+                    async move {
+                        let foo_val = i;
+                        let bar = format!("bar{}", i);
+
+                        let mut pipe = redis::pipe();
+                        pipe.atomic()
+                            .cmd("SET")
+                            .arg("key")
+                            .arg(foo_val)
+                            .ignore()
+                            .cmd("SET")
+                            .arg(&["key2", &bar[..]])
+                            .ignore()
+                            .cmd("MGET")
+                            .arg(&["key", "key2"]);
+
+                        pipe.query_async(&mut con)
+                            .map(move |result| {
+                                assert_eq!(Ok(((foo_val, bar.clone().into_bytes()),)), result);
+                                result
+                            })
+                            .await
+                    }
+                });
+                future::try_join_all(cmds)
+            })
+            .map_ok(|results| {
+                assert_eq!(results.len(), 100);
+            })
+            .map_err(|err| panic!("{}", err))
+            .await
+    })
     .unwrap();
 }
 
@@ -233,44 +244,42 @@ fn test_script() {
 
     let ctx = TestContext::new();
 
-    block_on_all(future::lazy(|| {
-        ctx.shared_async_connection().and_then(|con| {
-            script1
-                .key("key1")
-                .arg("foo")
-                .invoke_async(con)
-                .and_then(|(con, ())| script2.key("key1").invoke_async(con))
-                .and_then(|(con, val): (SharedConnection, String)| {
-                    assert_eq!(val, "foo");
-                    script1.key("key1").arg("bar").invoke_async(con)
-                })
-                .and_then(|(con, ())| script2.key("key1").invoke_async(con))
-                .and_then(|(_con, val): (SharedConnection, String)| {
-                    assert_eq!(val, "bar");
-                    Ok(())
-                })
-        })
-    }))
+    block_on_all(async move {
+        let mut con = ctx.multiplexed_async_connection().await?;
+        let () = script1
+            .key("key1")
+            .arg("foo")
+            .invoke_async(&mut con)
+            .await?;
+        let val: String = script2.key("key1").invoke_async(&mut con).await?;
+        assert_eq!(val, "foo");
+        let () = script1
+            .key("key1")
+            .arg("bar")
+            .invoke_async(&mut con)
+            .await?;
+        let val: String = script2.key("key1").invoke_async(&mut con).await?;
+        assert_eq!(val, "bar");
+        Ok(())
+    })
+    .map_err(|err: RedisError| err)
     .unwrap();
 }
 
 #[test]
 fn test_script_returning_complex_type() {
     let ctx = TestContext::new();
-    block_on_all(future::lazy(|| {
-        ctx.shared_async_connection().and_then(|con| {
-            redis::Script::new("return {1, ARGV[1], true}")
-                .arg("hello")
-                .invoke_async(con)
-                .and_then(
-                    |(_con, (i, s, b)): (SharedConnection, (i32, String, bool))| {
-                        assert_eq!(i, 1);
-                        assert_eq!(s, "hello");
-                        assert_eq!(b, true);
-                        Ok(())
-                    },
-                )
-        })
-    }))
+    block_on_all(async {
+        let mut con = ctx.multiplexed_async_connection().await?;
+        redis::Script::new("return {1, ARGV[1], true}")
+            .arg("hello")
+            .invoke_async(&mut con)
+            .map_ok(|(i, s, b): (i32, String, bool)| {
+                assert_eq!(i, 1);
+                assert_eq!(s, "hello");
+                assert_eq!(b, true);
+            })
+            .await
+    })
     .unwrap();
 }